🎁 fix: insert decimal number in recipe form page

+++ features/recipes/pages/recipe_form_page.tsx
@@ -393,7 +393,7 @@ export const RecipeFormPage = () => {
                                             formik.setFieldTouched("quantity");
                                             formik.setFieldValue(
                                               `quantity[${index}]`,
-                                              parseInt(e.target.value)
+                                              parseFloat(e.target.value)
                                             );
                                           }}
                                           type="number"
